fix(eink-dither): non-degeneracy assertions for pinning tests, clean clippy
Addresses Task 2 review findings:

- a_near_miss_is_not_pinned, plain_dither_is_unchanged_by_this_feature and
  pinning_is_refused_when_resizing asserted only with==without, relying on
  an out-of-band doc comment (not an assertion) that the scene was hostile.
  A future change that made the scene non-hostile (field colour, kernel,
  palette, clamp default) would collapse all three to tautologies with no
  failure and no signal. Each now asserts its own unpinned baseline is
  actually eroded (<0.99 ink share) before the equality — verified by
  perturbing the shared fixture to a non-hostile one (field == line) and
  confirming all five hostility-dependent tests now fail loudly on the
  precondition instead of silently passing the equality.

- Extracted the previously five-times-duplicated hostile-field fixture into
  hostile_line_field()/line_ink_share() so the strengthened assertion is a
  three-line addition at each site instead of another copy of the geometry,
  and so the scenario can't drift apart between tests.

- pinning_is_refused_when_resizing's doc comment overstated coverage: with
  .resize(w,h) equal to the input's own size (required so resize_lanczos's
  no-op branch avoids the missing-image-crate panic), the test verifies
  refusal is triggered by resize being *configured*, not that refusal
  prevents index misalignment from an actual dimension change — that
  scenario is untestable in this vendored build (no resampling backend).
  Doc comment now says so plainly instead of claiming both counts fixed.

- Cleared a `-D warnings`-failing clippy lint (casting to the same type is
  unnecessary) in dither/mod.rs, Task 1's file — not cosmetic, since it
  fails the workspace clippy gate `make check` runs.

Mutation-verified: a "kill the pin path entirely" mutant (pin_map forced
None) is still caught by the two positive tests
(eligibility_decides_where_pinning_applies,
the_exact_match_is_against_the_nominal_entry); the three negative-property
tests pass under it by construction, since "never pin" is consistent with
what they assert — expected, not a gap, given the positive tests already
cover that mutant.

+++ crates/eink-dither/src/dither/mod.rs
@@ -789,7 +789,7 @@ mod tests {
         // hops over it and the isolation claim is void.
         let max_dx = kernel.entries.iter().map(|&(dx, _, _)| dx).max().unwrap();
         assert!(
-            (BAR.end - BAR.start) as i32 > max_dx as i32,
+            (BAR.end - BAR.start) as i32 > max_dx,
             "pinned bar ({}px) is not wider than the kernel reach ({max_dx})",
             BAR.end - BAR.start
         );
